requirements.txt
click==8.1.7
GitPython==3.1.43
PyYAML==6.0.2
fastapi==0.115.5
uvicorn==0.32.0
httpx==0.27.2
jinja2==3.1.4
python-multipart==0.0.17
